Only load modules with ".el" suffix, not e.g. ".el~"
Normally this is not a problem because this config moves backup files
out of the modules directory to a central location. However, in
scenarios where for whatever reason that config step has not yet been
made, module file backups' filenames end with a ".el~" and get found
by this statement, causing all sorts of odd errors.